// dashboard/src/utils/hasura-api/generated/schemas/relationshipUsingForeignKeyConstraintOnForeignKeyConstraintOnOneOf.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { QualifiedTable } from './qualifiedTable';

/**
 * Column from remote table
 */
export type RelationshipUsingForeignKeyConstraintOnForeignKeyConstraintOnOneOf =
  {
    table?: QualifiedTable;
    columns?: string;
  };